+++ src/js/module/popup/comment.js
@@ -1,5 +1,4 @@
-import { TvAPI, baseUrl, appId } from '../API/API.js';
-import openPopup from './popup.js';
+import { TvAPI } from '../API/API.js';
 
 const generateCardHtml = (data) => `
   <div class="card col mt-3">
@@ -11,65 +10,4 @@ const generateCardHtml = (data) => `
   </div>
 `;
 
-const attachCommentButtonListeners = () => {
-  const commentButtons = document.querySelectorAll('.comment-button');
-  commentButtons.forEach((button) => {
-    button.addEventListener('click', async () => {
-      const tvApi = button.getAttribute('data-tvapi');
-      const response = await fetch(tvApi);
-      if (!response.ok) {
-        throw new Error(`Request failed with status ${response.status}`);
-      }
-      const data = await response.json();
-      openPopup(data);
-    });
-  });
-};
-
-const getReaction = async () => {
-  const url = `${baseUrl}${appId}/likes`;
-  const result = await fetch(`${url}`);
-  const data = await result.json();
-  return data;
-};
-
-const displayReaction = async (likeCounts) => {
-  const totalLikes = await getReaction();
-  totalLikes.forEach((totalLike) => {
-    likeCounts.forEach((likeCount) => {
-      if (totalLike.item_id === likeCount.id) {
-        likeCount.textContent = `${totalLike.likes} likes`;
-      }
-    });
-  });
-};
-
-const sendReactionToApi = async (likeBtn, likeCounts) => {
-  likeBtn.addEventListener('click', async (e) => {
-    const reactions = { item_id: `${e.target.id}` };
-    const url = `${baseUrl}${appId}/likes`;
-    const requestOptions = {
-      method: 'POST',
-      headers: {
-        'content-type': 'application/json; charset=UTF-8',
-      },
-      body: JSON.stringify(reactions),
-    };
-
-    await fetch(`${url}`, requestOptions);
-    displayReaction(likeCounts);
-  });
-};
-
-const attachLikeButtonListeners = () => {
-  const likeButtons = document.querySelectorAll('.like-button');
-  const likeCounts = document.querySelectorAll('.like-count');
-  likeButtons.forEach((likeButton) => sendReactionToApi(likeButton, likeCounts));
-  displayReaction(likeCounts);
-};
-
-export {
-  generateCardHtml,
-  attachCommentButtonListeners,
-  attachLikeButtonListeners,
-};
+export default generateCardHtml;
